Added file-statistics-implementation to CoreAccess/ProjectsManagingService
git-svn-id: svn://svn.jakeapp.com/trunk@1080 cfed719c-3465-4289-a5de-482d64f03f8e

+++ gui/src/main/java/com/jakeapp/gui/swing/CoreAccessMock.java
@@ -303,7 +303,7 @@ public class CoreAccessMock implements ICoreAccess {
 		return 100;
 	}
 
-	public int getProjectSizeTotal(Project project) {
+	public long getProjectSizeTotal(Project project) {
 		return 50000;
 	}
